% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPMon.R
\docType{package}
\name{FPMon}
\alias{FPMon}
\alias{FPMon-package}
\title{FPMon: a package sequential construction of parsers
Anton Antonov
Version 0.7}
\description{
FPMon: a package sequential construction of parsers
Anton Antonov
Version 0.7
}
