% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\docType{data}
\name{\%@>\%}
\alias{\%@>\%}
\title{Infix function for ParseApply.}
\format{An object of class \code{character} of length 1.}
\usage{
%@>%
}
\description{
Infix function for ParseApply.
}
\keyword{datasets}
