% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{ListToString}
\alias{ListToString}
\title{Convert a list into a pretty string.}
\usage{
ListToString(arg, leftBracket = "{", rightBracket = "}",
  headLeftBracket = NULL, headRightBracket = NULL, head = NULL,
  level = 0)
}
\arguments{
\item{arg}{A list.}

\item{leftBracket}{The symbol for the left bracket.}

\item{rightBracket}{The symbol for the right bracket.}

\item{headLeftBracket}{The "head" to be used for the left bracket.}

\item{headRightBracket}{The "head" to be used for the right bracket.}

\item{head}{A head symbol.}

\item{level}{A integer used in recursion.}
}
\value{
A string.
}
\description{
Convert a list into a pretty string.
}
\details{
This function is intended to give Mathematica-lists looking results.
}
