% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{p.alt}
\alias{p.alt}
\title{Synonym of \code{ParseAlternativeComposition}.}
\usage{
p.alt(parsers)
}
\description{
Synonym of \code{ParseAlternativeComposition}.
}
\seealso{
\code{\link{ParseAlternativeComposition}} is the original function.

Other "short names" parsers: \code{\link{p.apply}},
  \code{\link{p.bracketed}}, \code{\link{p.chainl}},
  \code{\link{p.chainr}}, \code{\link{p.curlybracketed}},
  \code{\link{p.fail}}, \code{\link{p.just}},
  \code{\link{p.listof}}, \code{\link{p.many1}},
  \code{\link{p.many}}, \code{\link{p.option}},
  \code{\link{p.pack}}, \code{\link{p.parenthesized}},
  \code{\link{p.pred}}, \code{\link{p.satisfy}},
  \code{\link{p.seql}}, \code{\link{p.seqr}},
  \code{\link{p.seq}}, \code{\link{p.shortest}},
  \code{\link{p.some}}, \code{\link{p.symbol}}
}
\concept{"short names" parsers}
