% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseAlternativeComposition}
\alias{ParseAlternativeComposition}
\title{Compose parsers for alternate parsing.}
\usage{
ParseAlternativeComposition(parsers)
}
\arguments{
\item{parsers}{A list of parsers.}
}
\value{
A function that is a parser.
}
\description{
Gives a parser that is an alternating composition of parsers.
}
\seealso{
\code{\link{p.alt}} is a shorter synonym;
  \code{\link{ParseSequentialComposition}} for sequential composition.

Other basic combinators: \code{\link{ParseSequentialComposition}}
}
\concept{basic combinators}
