% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseApply}
\alias{ParseApply}
\title{Apply a function to parsing results.}
\usage{
ParseApply(f, p)
}
\arguments{
\item{f}{A function to applied to the parsing results.}

\item{p}{A parser.}
}
\value{
A function that is a parser.
}
\description{
Gives a parser that modifies the parsing results with specified function.
}
\seealso{
Other next combinators: \code{\link{ParseJust}},
  \code{\link{ParseSequentialCompositionPickLeft}},
  \code{\link{ParseSequentialCompositionPickRight}},
  \code{\link{ParseShortest}}, \code{\link{ParseSome}}
}
\concept{next combinators}
