% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseBracketed}
\alias{ParseBracketed}
\title{Parsing token(s) between brackets.}
\usage{
ParseBracketed(p)
}
\arguments{
\item{p}{A parser of tokens between brackets}
}
\value{
A function that is a parser.
}
\description{
Used to parse rules that adhere to the pattern "[ expr ]".
}
\seealso{
Other Pack parsers.: \code{\link{ParseCurlyBracketed}},
  \code{\link{ParsePack}}, \code{\link{ParseParenthesized}}
}
\concept{Pack parsers.}
