% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseChainLeft}
\alias{ParseChainLeft}
\title{Left parsing a chain.}
\usage{
ParseChainLeft(p, sp)
}
\arguments{
\item{p}{A parser.}

\item{sepParser}{A separator parser.}
}
\value{
A function that is a parser.
}
\description{
Parsing a chain over a certain operation.
}
\seealso{
Other Next parsers.: \code{\link{ParseChainRight}},
  \code{\link{ParseChain}}, \code{\link{ParseListOf}},
  \code{\link{ParseMany1}}, \code{\link{ParseMany}},
  \code{\link{ParseOption1}}, \code{\link{ParseOption}}
}
\concept{Next parsers.}
