% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseComposeWithResults}
\alias{ParseComposeWithResults}
\title{Composer of parsing results.}
\usage{
ParseComposeWithResults(p, res)
}
\arguments{
\item{p}{A parser to be applied to the parsing results.}

\item{res}{A list of parsing results.}
}
\value{
A list of parsing results.
}
\description{
Gives a list of the results of a parser applied over a list.
}
