% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseCurlyBracketed}
\alias{ParseCurlyBracketed}
\title{Parsing token(s) between curly brackets.}
\usage{
ParseCurlyBracketed(p)
}
\arguments{
\item{p}{A parser of tokens between curly brackets}
}
\value{
A function that is a parser.
}
\description{
Used to parse rules that adhere to the pattern "{ expr }".
}
\seealso{
Other Pack parsers.: \code{\link{ParseBracketed}},
  \code{\link{ParsePack}}, \code{\link{ParseParenthesized}}
}
\concept{Pack parsers.}
