% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseEpsilon}
\alias{ParseEpsilon}
\title{Parser that does nothing.}
\usage{
ParseEpsilon(xs)
}
\value{
A function that is a parser.
}
\description{
Gives a parser that parses nothing.
}
\seealso{
Other basic parsers: \code{\link{ParseFailed}},
  \code{\link{ParsePredicate}}, \code{\link{ParseSucceed}},
  \code{\link{ParseSymbol}}
}
\concept{basic parsers}
