% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseFailed}
\alias{ParseFailed}
\title{Parser that fails.}
\usage{
ParseFailed(xs)
}
\value{
A function that is a parser.
}
\description{
Gives a parser that returns NULL.
}
\seealso{
Other basic parsers: \code{\link{ParseEpsilon}},
  \code{\link{ParsePredicate}}, \code{\link{ParseSucceed}},
  \code{\link{ParseSymbol}}
}
\concept{basic parsers}
