% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseJust}
\alias{ParseJust}
\title{Take parsing results with no residuals.}
\usage{
ParseJust(p)
}
\arguments{
\item{p}{A parser.}
}
\value{
A function that is a parser.
}
\description{
Gives a parser that takes only completely successful results.
}
\seealso{
Other next combinators: \code{\link{ParseApply}},
  \code{\link{ParseSequentialCompositionPickLeft}},
  \code{\link{ParseSequentialCompositionPickRight}},
  \code{\link{ParseShortest}}, \code{\link{ParseSome}}
}
\concept{next combinators}
