% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseListOf}
\alias{ParseListOf}
\title{Parsing a list of many.}
\usage{
ParseListOf(p, sepParser)
}
\arguments{
\item{p}{A parser.}

\item{sepParser}{A separator parser.}
}
\value{
A function that is a parser.
}
\description{
Used to parse a sequence of many patterns corresponding of a parser 
  separated by patterns corresponding to a second parser.
}
\seealso{
Other Next parsers.: \code{\link{ParseChainLeft}},
  \code{\link{ParseChainRight}}, \code{\link{ParseChain}},
  \code{\link{ParseMany1}}, \code{\link{ParseMany}},
  \code{\link{ParseOption1}}, \code{\link{ParseOption}}
}
\concept{Next parsers.}
