% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseMany}
\alias{ParseMany}
\title{Parsing a sequence of many.}
\usage{
ParseMany(p)
}
\arguments{
\item{p}{A parser.}
}
\value{
A function that is a parser.
}
\description{
Used to parse a sequence of many patterns corresponding of a parser.
}
\seealso{
Other Next parsers.: \code{\link{ParseChainLeft}},
  \code{\link{ParseChainRight}}, \code{\link{ParseChain}},
  \code{\link{ParseListOf}}, \code{\link{ParseMany1}},
  \code{\link{ParseOption1}}, \code{\link{ParseOption}}
}
\concept{Next parsers.}
