% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseOption1}
\alias{ParseOption1}
\title{Parsing optionally (second version).}
\usage{
ParseOption1(p)
}
\arguments{
\item{p}{A parser.}
}
\value{
A function that is a parser.
}
\description{
Used to parse optional presence.
}
\seealso{
Other Next parsers.: \code{\link{ParseChainLeft}},
  \code{\link{ParseChainRight}}, \code{\link{ParseChain}},
  \code{\link{ParseListOf}}, \code{\link{ParseMany1}},
  \code{\link{ParseMany}}, \code{\link{ParseOption}}
}
\concept{Next parsers.}
