% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParsePack}
\alias{ParsePack}
\title{Parsing a triplet-pack.}
\usage{
ParsePack(s1, p, s2)
}
\arguments{
\item{s1}{A parser (left separator).}

\item{p}{A parser (middle).}

\item{s2}{A parser (right separator).}
}
\value{
A function that is a parser.
}
\description{
Used to parse rules that adhere to the pattern "separator1 expr separator2".
}
\seealso{
Other Pack parsers.: \code{\link{ParseBracketed}},
  \code{\link{ParseCurlyBracketed}},
  \code{\link{ParseParenthesized}}
}
\concept{Pack parsers.}
