% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseParenthesized}
\alias{ParseParenthesized}
\title{Parsing token(s) between parentheses.}
\usage{
ParseParenthesized(p)
}
\arguments{
\item{p}{A parser of tokens between parentheses.}
}
\value{
A function that is a parser.
}
\description{
Used to parse rules that adhere to the pattern "( parser )".
}
\seealso{
Other Pack parsers.: \code{\link{ParseBracketed}},
  \code{\link{ParseCurlyBracketed}},
  \code{\link{ParsePack}}
}
\concept{Pack parsers.}
