% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParsePredicate}
\alias{ParsePredicate}
\title{Parsing with a predicate.}
\usage{
ParsePredicate(predFunc)
}
\arguments{
\item{predFunc}{A function that is a certain predicate.}
}
\value{
A function that is a parser.
}
\description{
Gives a parser for a given predicate.
}
\seealso{
Other basic parsers: \code{\link{ParseEpsilon}},
  \code{\link{ParseFailed}}, \code{\link{ParseSucceed}},
  \code{\link{ParseSymbol}}
}
\concept{basic parsers}
