% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseSequentialComposition}
\alias{ParseSequentialComposition}
\title{Compose parsers for sequential parsing.}
\usage{
ParseSequentialComposition(parsers)
}
\arguments{
\item{parsers}{A list of parsers.}
}
\value{
A function that is a parser.
}
\description{
Gives a parser that is a sequential composition of parsers.
}
\details{
A basic combinator of parsers. Note the argument is a list of parsers.
  A (probably) better version would take "just" arguments.
  \code{\alias{\%&\%}} is an infix version;
}
\seealso{
\code{\link{p.seq}} is a shorter synonym;
  \code{\link{ParseAlternativeComposition}} for alternating composition;
  \code{\link{ParseComposeWithResults}} for how the results are obtained.

Other basic combinators: \code{\link{ParseAlternativeComposition}}
}
\concept{basic combinators}
