% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseSequentialCompositionPickLeft}
\alias{ParseSequentialCompositionPickLeft}
\title{Take the left parsing result.}
\usage{
ParseSequentialCompositionPickLeft(p1, p2)
}
\arguments{
\item{p}{A parser.}
}
\value{
A function that is a parser.
}
\description{
Gives a parser that takes the left parsing result 
  of a sequential composition of two parsers.
}
\details{
Note that the function takes two parsers not a list of parsers
  as \code{\link{ParseSequentialComposition}}.
}
\seealso{
\code{\link{ParseSequentialComposition}} is the generic function;
  \code{\link{ParseSequentialCompositionPickRight}} for picking the right result.

Other next combinators: \code{\link{ParseApply}},
  \code{\link{ParseJust}},
  \code{\link{ParseSequentialCompositionPickRight}},
  \code{\link{ParseShortest}}, \code{\link{ParseSome}}
}
\concept{next combinators}
