% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseSequentialCompositionPickRight}
\alias{ParseSequentialCompositionPickRight}
\title{Take the right parsing result.}
\usage{
ParseSequentialCompositionPickRight(p1, p2)
}
\arguments{
\item{p}{A parser.}
}
\value{
A function that is a parser.
}
\description{
Gives a parser that takes the right parsing result 
  of a sequential composition of two parsers.
}
\details{
Note that the function takes two parsers not a list of parsers
  as \code{\link{ParseSequentialComposition}}.
}
\seealso{
\code{\link{ParseSequentialComposition}} is the generic function;
  \code{\link{ParseSequentialCompositionPickLeft}} for picking the left result.

Other next combinators: \code{\link{ParseApply}},
  \code{\link{ParseJust}},
  \code{\link{ParseSequentialCompositionPickLeft}},
  \code{\link{ParseShortest}}, \code{\link{ParseSome}}
}
\concept{next combinators}
