% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseShortest}
\alias{ParseShortest}
\title{Take the parsing result with the shortest residual.}
\usage{
ParseShortest(p)
}
\arguments{
\item{p}{A parser.}
}
\value{
A function that is a parser.
}
\description{
Gives a parser that takes the result with the shortest residual.
}
\seealso{
Other next combinators: \code{\link{ParseApply}},
  \code{\link{ParseJust}},
  \code{\link{ParseSequentialCompositionPickLeft}},
  \code{\link{ParseSequentialCompositionPickRight}},
  \code{\link{ParseSome}}
}
\concept{next combinators}
