% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseSome}
\alias{ParseSome}
\title{Take the first completely successful parsing result.}
\usage{
ParseSome(p)
}
\arguments{
\item{p}{A parser.}
}
\value{
A function that is a parser.
}
\description{
Gives a parser that takes the first result of \code{\link{ParseJust}}.
}
\seealso{
\code{\link{ParseJust}} has similar functionality.

Other next combinators: \code{\link{ParseApply}},
  \code{\link{ParseJust}},
  \code{\link{ParseSequentialCompositionPickLeft}},
  \code{\link{ParseSequentialCompositionPickRight}},
  \code{\link{ParseShortest}}
}
\concept{next combinators}
