% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseSucceed}
\alias{ParseSucceed}
\title{Parser that succeeds.}
\usage{
ParseSucceed(v)
}
\arguments{
\item{v}{A "symbol" that is the successful parsing result.}
}
\value{
A function that is a parser.
}
\description{
Gives a parser that does a successful parsing.
}
\seealso{
Other basic parsers: \code{\link{ParseEpsilon}},
  \code{\link{ParseFailed}}, \code{\link{ParsePredicate}},
  \code{\link{ParseSymbol}}
}
\concept{basic parsers}
