% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalParsers.R
\name{ParseSymbol}
\alias{ParseSymbol}
\title{Parsing a "symbol".}
\usage{
ParseSymbol(a)
}
\arguments{
\item{a}{A string.}
}
\value{
A function that is a parser.
}
\description{
Gives a parser for a given string.
}
\seealso{
Other basic parsers: \code{\link{ParseEpsilon}},
  \code{\link{ParseFailed}}, \code{\link{ParsePredicate}},
  \code{\link{ParseSucceed}}
}
\concept{basic parsers}
